moved versionControl to controls folder
